Cube fixes in progress. 2d matrix and 3d cube now both get altered when moving a face
