+++ src/animated/Cube.tsx
@@ -1,89 +1,12 @@
 import React, { useRef, useEffect } from 'react';
 import * as THREE from 'three';
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
-import { useRubiksCube } from './RubiksCubeContext';
 
 const Cube: React.FC = () => {
-    const { rubiksCubeMatrix, setRubiksCube } = useRubiksCube();
+
     const mountRef = useRef<HTMLDivElement | null>(null);
 
-
-    const rotateFaceCounterclockwise = (matrix: string[][]): string[][] => {
-        const rotated = matrix[0].map((_, index) =>
-            matrix.map(row => row[index]).reverse()
-        );
-        console.log('Matrix after clockwise rotation:', rotated);
-        return rotated;
-    };
-
-    const rotateFaceClockwise = (matrix: string[][]): string[][] => {
-        const rotated = matrix[0].map((_, index) =>
-            matrix.map(row => row[matrix.length - 1 - index])
-        );
-        console.log('Matrix after counterclockwise rotation:', rotated);
-        return rotated;
-    };
-
-    const scrambleMatrix = () => {
-        const axes = ['x', 'y', 'z'] as const;
-        const directions = [-1, 1];
-        const layers = [-1, 0, 1];
-
-        // Number of random moves to scramble
-        const scrambleMoves = 20;
-
-        for (let i = 0; i < scrambleMoves; i++) {
-            const axis = axes[Math.floor(Math.random() * axes.length)];
-            const direction = directions[Math.floor(Math.random() * directions.length)];
-            const layer = layers[Math.floor(Math.random() * layers.length)];
-
-            const angle = direction * Math.PI / 2;
-
-
-            // Update the logical matrix
-            if (axis === 'x') {
-                if (layer === -1) rubiksCubeMatrix.left = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.left)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.left);
-                if (layer === 1) rubiksCubeMatrix.right = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.right)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.right);
-            } else if (axis === 'y') {
-                if (layer === -1) rubiksCubeMatrix.bottom = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.bottom)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.bottom);
-                if (layer === 1) rubiksCubeMatrix.top = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.top)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.top);
-            } else if (axis === 'z') {
-                if (layer === -1) rubiksCubeMatrix.back = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.back)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.back);
-                if (layer === 1) rubiksCubeMatrix.front = direction === 1
-                    ? rotateFaceClockwise(rubiksCubeMatrix.front)
-                    : rotateFaceCounterclockwise(rubiksCubeMatrix.front);
-            }
-        }
-
-        // Update the Cube state
-        setRubiksCube({ ...rubiksCubeMatrix });
-
-
-    };
-
-
-
-
-    // End of matrix handling
-
-
-
-
-
-
     useEffect(() => {
-
-
         const scene = new THREE.Scene();
         const camera = new THREE.PerspectiveCamera(50, window.innerWidth / window.innerHeight, 0.1, 1000);
         camera.position.z = 8;
@@ -124,12 +47,12 @@ const Cube: React.FC = () => {
 
                     // Assign colors to faces based on position
                     const materials = [
-                        new THREE.MeshBasicMaterial({color: x === 1 ? 0x0000ff : 0x000000}), // Right: Blue
-                        new THREE.MeshBasicMaterial({color: x === -1 ? 0x00ff00 : 0x000000}), // Left: Green
-                        new THREE.MeshBasicMaterial({color: y === 1 ? 0xff0000 : 0x000000}), // Top: Red
-                        new THREE.MeshBasicMaterial({color: y === -1 ? 0xffa500 : 0x000000}), // Bottom: Orange
-                        new THREE.MeshBasicMaterial({color: z === 1 ? 0xffffff : 0x000000}), // Front: gray
-                        new THREE.MeshBasicMaterial({color: z === -1 ? 0xffff00 : 0x000000}), // Back: Yellow
+                        new THREE.MeshBasicMaterial({color: x === 1 ? 0x0000ff : 0x000000}),
+                        new THREE.MeshBasicMaterial({color: x === -1 ? 0x00ff00 : 0x000000}),
+                        new THREE.MeshBasicMaterial({color: y === 1 ? 0xff0000 : 0x000000}),
+                        new THREE.MeshBasicMaterial({color: y === -1 ? 0xffa500 : 0x000000}),
+                        new THREE.MeshBasicMaterial({color: z === 1 ? 0xffffff : 0x000000}),
+                        new THREE.MeshBasicMaterial({color: z === -1 ? 0xffff00 : 0x000000}),
                     ];
 
                     const cube = new THREE.Mesh(geometry, materials);
@@ -139,7 +62,6 @@ const Cube: React.FC = () => {
                 }
             }
         }
-
 
         // Helper function to get cubes on a specific face
         const getFaceCubes = (axis: 'x' | 'y' | 'z', index: number) => {
@@ -159,8 +81,8 @@ const Cube: React.FC = () => {
 
             // Attach the cubes to the pivot for rotation
             faceCubes.forEach(cube => {
-                cubeGroup.remove(cube); // Remove from the main group
-                pivot.add(cube); // Add to the pivot group
+                cubeGroup.remove(cube);
+                pivot.add(cube);
             });
 
             // Rotate the pivot group
@@ -176,257 +98,64 @@ const Cube: React.FC = () => {
             pivot.updateMatrixWorld();
             faceCubes.forEach(cube => {
                 cube.applyMatrix4(pivot.matrixWorld);
-                pivot.remove(cube); // Remove from the pivot group
-                cubeGroup.add(cube); // Reattach to the main group
+
+                // Remove from the pivot group
+                pivot.remove(cube);
+
+                // Reattach to the main group
+                cubeGroup.add(cube);
             });
 
-            scene.remove(pivot); // Remove the pivot from the scene
+            // Remove the pivot
+            scene.remove(pivot);
         };
 
-        // Keydown event handler to rotate faces
         const handleKeyDown = (event: KeyboardEvent) => {
-
             switch (event.key) {
-                // Front face (Z=1)
-                case 'v': // Front clockwise
-                {
 
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    newRubiksCubeMatrix.front = rotateFaceClockwise(newRubiksCubeMatrix.front);
-
-
-                    const topRowV = [...newRubiksCubeMatrix.top[2]];
-                    const rightColV = newRubiksCubeMatrix.right.map(row => row[0]);
-                    const bottomRowV = [...newRubiksCubeMatrix.bottom[0]];
-                    const leftColV = newRubiksCubeMatrix.left.map(row => row[2]);
-
-                    newRubiksCubeMatrix.top[2] = leftColV.reverse();
-                    newRubiksCubeMatrix.right.forEach((row, i) => (row[0] = topRowV[topRowV.length - 1 - i]));
-                    newRubiksCubeMatrix.bottom[0] = rightColV.reverse();
-                    newRubiksCubeMatrix.left.forEach((row, i) => (row[2] = bottomRowV[i]));
-
-                    setRubiksCube(newRubiksCubeMatrix); // Ensure matrix is updated
+                // Front rotate clockwise
+                case 'v':
+                    rotateFace('z', 1, -Math.PI / 2);
                     break;
-                }
 
-
-
-
-
-                // Back face (Z=-1)
-                case 'b': // Back clockwise
-                {
-
-                    rotateFace('z', -1, -Math.PI / 2);
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    newRubiksCubeMatrix.back = rotateFaceClockwise(newRubiksCubeMatrix.back);
-
-                    const topRow = [...newRubiksCubeMatrix.top[0]];
-                    const rightCol = newRubiksCubeMatrix.right.map((row: string[]) => row[2]);
-                    const bottomRow = [...newRubiksCubeMatrix.bottom[2]];
-                    const leftCol = newRubiksCubeMatrix.left.map((row: string[]) => row[0]);
-
-                    newRubiksCubeMatrix.top[0] = leftCol.reverse();
-                    newRubiksCubeMatrix.right.forEach((row: string[], i: number) => (row[2] = topRow[i]));
-                    newRubiksCubeMatrix.bottom[2] = rightCol.reverse();
-                    newRubiksCubeMatrix.left.forEach((row: string[], i: number) => (row[0] = bottomRow[i]));
-                    setRubiksCube(newRubiksCubeMatrix)
-
-                    break;
-                }
-
-                case 'n': // Back counterclockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    rotateFace('z', -1, Math.PI / 2);
-                    newRubiksCubeMatrix.back = rotateFaceCounterclockwise(newRubiksCubeMatrix.back);
-
-                    const topRow = [...newRubiksCubeMatrix.top[0]];
-                    const rightCol = newRubiksCubeMatrix.right.map((row: string[]) => row[2]);
-                    const bottomRow = [...newRubiksCubeMatrix.bottom[2]];
-                    const leftCol = newRubiksCubeMatrix.left.map((row: string[]) => row[0]);
-
-
-                    newRubiksCubeMatrix.top[0] = rightCol.reverse();
-                    newRubiksCubeMatrix.right.forEach((row: string[], i: number) => (row[2] = bottomRow[i]));
-                    newRubiksCubeMatrix.bottom[2] = leftCol.reverse();
-                    newRubiksCubeMatrix.left.forEach((row: string[], i: number) => (row[0] = topRow[i]));
-                    setRubiksCube(newRubiksCubeMatrix);
-
-
-
-                    break;
-                }
-
-                case 'f': // Front counterclockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
+                // Front counterclockwise
+                case 'f':
                     rotateFace('z', 1, Math.PI / 2);
-                    newRubiksCubeMatrix.front = rotateFaceCounterclockwise(newRubiksCubeMatrix.front);
-
-                    const topRow = [...newRubiksCubeMatrix.top[2]];
-                    const rightCol = newRubiksCubeMatrix.right.map((row: string[]) => row[0]);
-                    const bottomRow = [...newRubiksCubeMatrix.bottom[0]];
-                    const leftCol = newRubiksCubeMatrix.left.map((row: string[]) => row[2]);
-
-                    newRubiksCubeMatrix.top[2] = rightCol;
-                    newRubiksCubeMatrix.right.forEach((row: string[], i: number) => (row[0] = bottomRow[i]));
-                    newRubiksCubeMatrix.bottom[0] = leftCol.reverse();
-                    newRubiksCubeMatrix.left.forEach((row: string[], i: number) => (row[2] = topRow[i]));
-                    setRubiksCube(newRubiksCubeMatrix)
-
                     break;
-                }
+
+                // Back clockwise
+                case 'b':
+                    rotateFace('z', -1, -Math.PI / 2);
+                    break;
+
+                // Back counterclockwise
+                case 'n':
+                    rotateFace('z', -1, Math.PI / 2);
+                    break;
+
+                // Left clockwise
+                case 'l':
+                    rotateFace('x', -1, -Math.PI / 2);
+                    break;
+
+                // Left counterclockwise
+                case 'k':
+                    rotateFace('x', -1, Math.PI / 2);
+                    break;
+
+                // Right clockwise
+                case 'r':
+                    rotateFace('x', 1, -Math.PI / 2);
+                    break;
 
                 case 't': // Right counterclockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
                     rotateFace('x', 1, Math.PI / 2);
-                    newRubiksCubeMatrix.right = rotateFaceCounterclockwise(newRubiksCubeMatrix.right);
-
-                    const topCol = newRubiksCubeMatrix.top.map((row: string[]) => row[2]);
-                    const frontCol = newRubiksCubeMatrix.front.map((row: string[]) => row[2]);
-                    const bottomCol = newRubiksCubeMatrix.bottom.map((row: string[]) => row[2]);
-                    const backCol = newRubiksCubeMatrix.back.map((row: string[]) => row[0]).reverse();
-
-                    newRubiksCubeMatrix.top.forEach((row: string[], i: number) => (row[2] = backCol[i]));
-                    newRubiksCubeMatrix.front.forEach((row: string[], i: number) => (row[2] = topCol[i]));
-                    newRubiksCubeMatrix.bottom.forEach((row: string[], i: number) => (row[2] = frontCol[i]));
-                    newRubiksCubeMatrix.back.forEach((row: string[], i: number) => (row[0] = bottomCol[i]));
-                    setRubiksCube(newRubiksCubeMatrix);
-
                     break;
-                }
-                // Left face (X=-1)
-                case 'l': // Left clockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    rotateFace('x', -1, -Math.PI / 2);
-                    newRubiksCubeMatrix.left = rotateFaceClockwise(newRubiksCubeMatrix.left);
-
-                    const topCol = newRubiksCubeMatrix.top.map((row: string[]) => row[0]);
-                    const frontCol = newRubiksCubeMatrix.front.map((row: string[]) => row[0]);
-                    const bottomCol = newRubiksCubeMatrix.bottom.map((row: string[]) => row[0]);
-                    const backCol = newRubiksCubeMatrix.back.map((row: string[]) => row[2]).reverse();
-
-                    newRubiksCubeMatrix.top.forEach((row: string[], i: number) => (row[0] = frontCol[i]));
-                    newRubiksCubeMatrix.front.forEach((row: string[], i: number) => (row[0] = bottomCol[i]));
-                    newRubiksCubeMatrix.bottom.forEach((row: string[], i: number) => (row[0] = backCol[i]));
-                    newRubiksCubeMatrix.back.forEach((row: string[], i: number) => (row[2] = topCol[i]));
-                    setRubiksCube(newRubiksCubeMatrix);
-
-                    break;
-                }
-
-                case 'k': // Left counterclockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    rotateFace('x', -1, Math.PI / 2);
-                    newRubiksCubeMatrix.left = rotateFaceCounterclockwise(newRubiksCubeMatrix.left);
-
-                    const topCol = newRubiksCubeMatrix.top.map((row: string[]) => row[0]);
-                    const frontCol = newRubiksCubeMatrix.front.map((row: string[]) => row[0]);
-                    const bottomCol = newRubiksCubeMatrix.bottom.map((row: string[]) => row[0]);
-                    const backCol = newRubiksCubeMatrix.back.map((row: string[]) => row[2]).reverse();
-
-                    newRubiksCubeMatrix.top.forEach((row: string[], i: number) => (row[0] = backCol[i]));
-                    newRubiksCubeMatrix.front.forEach((row: string[], i: number) => (row[0] = topCol[i]));
-                    newRubiksCubeMatrix.bottom.forEach((row: string[], i: number) => (row[0] = frontCol[i]));
-                    newRubiksCubeMatrix.back.forEach((row: string[], i: number) => (row[2] = bottomCol[i]));
-                    setRubiksCube(newRubiksCubeMatrix);
-
-                    break;
-                }
-
-                case 'r': // Right clockwise
-                {
-                    const newRubiksCubeMatrix = { ...rubiksCubeMatrix };
-                    rotateFace('x', 1, -Math.PI / 2);
-                    newRubiksCubeMatrix.right = rotateFaceClockwise(newRubiksCubeMatrix.right);
-
-                    const topCol = newRubiksCubeMatrix.top.map((row: string[]) => row[2]);
-                    const frontCol = newRubiksCubeMatrix.front.map((row: string[]) => row[2]);
-                    const bottomCol = newRubiksCubeMatrix.bottom.map((row: string[]) => row[2]);
-                    const backCol = newRubiksCubeMatrix.back.map((row: string[]) => row[0]).reverse();
-
-                    newRubiksCubeMatrix.top.forEach((row: string[], i: number) => (row[2] = frontCol[i]));
-                    newRubiksCubeMatrix.front.forEach((row: string[], i: number) => (row[2] = bottomCol[i]));
-                    newRubiksCubeMatrix.bottom.forEach((row: string[], i: number) => (row[2] = backCol[i]));
-                    newRubiksCubeMatrix.back.forEach((row: string[], i: number) => (row[0] = topCol[i]));
-                    setRubiksCube(newRubiksCubeMatrix);
-
-                    break;
-                }
-
-
-
-                // Add cases for top (`u`, `i`) and bottom (`d`, `c`)...
-
-                case 'q': // Reset the matrix
-                    setRubiksCube({
-                        front: Array(3).fill(null).map(() => Array(3).fill('white')),
-                        back: Array(3).fill(null).map(() => Array(3).fill('yellow')),
-                        left: Array(3).fill(null).map(() => Array(3).fill('green')),
-                        right: Array(3).fill(null).map(() => Array(3).fill('blue')),
-                        top: Array(3).fill(null).map(() => Array(3).fill('red')),
-                        bottom: Array(3).fill(null).map(() => Array(3).fill('orange')),
-                    });
-                    break;
-
-                case 's': // Shuffle
-                {
-                    const shuffledRubiksCubeMatrix = {
-                        front: [
-                            ['green', 'white', 'blue'],
-                            ['orange', 'red', 'green'],
-                            ['blue', 'yellow', 'orange'],
-                        ],
-                        back: [
-                            ['yellow', 'orange', 'white'],
-                            ['green', 'red', 'yellow'],
-                            ['blue', 'red', 'white'],
-                        ],
-                        left: [
-                            ['red', 'green', 'yellow'],
-                            ['blue', 'orange', 'red'],
-                            ['white', 'yellow', 'green'],
-                        ],
-                        right: [
-                            ['yellow', 'blue', 'orange'],
-                            ['white', 'green', 'red'],
-                            ['yellow', 'orange', 'blue'],
-                        ],
-                        top: [
-                            ['red', 'yellow', 'blue'],
-                            ['orange', 'white', 'green'],
-                            ['green', 'blue', 'white'],
-                        ],
-                        bottom: [
-                            ['blue', 'red', 'yellow'],
-                            ['green', 'orange', 'white'],
-                            ['red', 'yellow', 'orange'],
-                        ],
-                    };
-
-                    setRubiksCube(shuffledRubiksCubeMatrix);
-
-                    break;
-                }
 
 
             }
 
-
-
-
-
         };
-
-        // Functions to alter matrix
-
-
-
-
 
         window.addEventListener('keydown', handleKeyDown);
 
@@ -453,7 +182,7 @@ const Cube: React.FC = () => {
                 mountRef.current.removeChild(renderer.domElement);
             }
         };
-    }, [rubiksCubeMatrix, setRubiksCube]);
+    }, []);
 
     return <div ref={mountRef} style={{ width: '100vw', height: '100vh' }}
     />;
